requirements.txt
# ============================================================
# Scientific Document Intelligence Pipeline — Requirements
# ============================================================

# ─── Core AI / OpenAI ─────────────────────────────────────────
openai>=1.40.0

# ─── Document Ingestion ───────────────────────────────────────
PyMuPDF>=1.24.0              # Fast PDF parsing (fitz)
pdfminer.six>=20221105       # PDF fallback parser
python-docx>=1.1.0           # Word document (.docx) parsing
trafilatura>=1.9.0            # Web URL article extraction (primary)
httpx>=0.27.0                 # Async-capable HTTP client for web loading
beautifulsoup4>=4.12.0       # HTML parsing fallback
lxml>=5.0.0                  # XML/HTML parser backend

# ─── Vector Store ─────────────────────────────────────────────
chromadb>=0.5.0              # ChromaDB client

# ─── Knowledge Graph ──────────────────────────────────────────
neo4j>=5.20.0                # Neo4j Python driver

# ─── Config & CLI ─────────────────────────────────────────────
pyyaml>=6.0.1                # YAML config file parsing
click>=8.1.0                 # CLI framework

# ─── Data Processing ──────────────────────────────────────────
numpy>=1.26.0                # Numerical ops

# ─── Development & Testing ────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# ─── Code Quality (optional, used by make lint/format) ────────
ruff>=0.4.0
mypy>=1.10.0
